refactor(smart-ptr): move codec traits to detail header
